fixed usb scripts, tweaked vim line numbers

+++ Scripts/umount-usb-drive.sh
@@ -1,7 +1,22 @@
 #!/bin/bash
 
-echo "Umounting drive from dir \"usb\" "
-umount  usb
-echo "Removing dir \"usb\" "
-rmdir /home/josh/usb
-echo "Done"
+echo -n "Are you sure you want to unmount \"/usb\" (y/n)? "
+
+read answer
+if echo "$answer" | grep -iq "^y" ; then
+    
+    echo "Attempting to unmount drive from dir \"/usb\""
+    if $(umount usb); then
+        echo "Unmount successful."
+        rmdir /home/josh/usb
+        exit 1
+    else
+        echo "Unmount failed, exiting"
+        exit 1
+    fi
+
+else
+    echo "Exiting"
+    exit 1
+fi
+
